+++ src/ctjhai/types.h
@@ -32,13 +32,13 @@ typedef enum { C_0MOD2 = 1, C_1MOD2, C_3MOD4, C_0MOD4, C_0MOD3 } mod_t;
 
 typedef struct {
 	mod_t weight_constraint;
-	unsigned short known_lower_bound;
-	unsigned short lower_bound, upper_bound;
-	unsigned short displayed_lower_bound;
-	unsigned short info_weight_completed;
-	unsigned short max_info_weight_req;
-	unsigned short ith_matrix;
-	unsigned short matrices_req;
+	unsigned int known_lower_bound;
+	unsigned int lower_bound, upper_bound;
+	unsigned int displayed_lower_bound;
+	unsigned int info_weight_completed;
+	unsigned int max_info_weight_req;
+	unsigned int ith_matrix;
+	unsigned int matrices_req;
 } INFO;
 /*-----------------------------------------------------------------*/
